Blog – 4 points
Create a page called News to display the Blog posts (create at least 3 posts).

On the News page show the excerpts of the posts instead of the full content and display the sidebar before the footer instead of next to the content.

Setup Animate on Scroll so each blog post animates into the viewport when scrolling. Include the Animate on Scroll CSS and JS files in your theme instead of using the CDN links. Only enqueue the files for the Post Type of ‘post’.

Home Page – 4 points

Display a static home page.

Add content to the block editor including at least one block with “Wide width” set and at least one block with “Full width” set. Add support for the Wide and Full alignment options in the Block Editor and add styles for both classes to complete this task. (Note: example code for this is in the fwd-starter-theme functions.php and style.css files.)

Display the title and featured image of the three most recent blog posts at the bottom of the home page and link them to their single post page.

Staff CPT & Taxonomy – 4 points

Register a Custom Post Type (CPT) for staff that only supports the title.

Register a custom taxonomy for the staff CPT and create the taxonomy terms Faculty and Administrative.

Add Advanced Custom Fields to the staff CPT to display the following information:

• Short staff biography

• Course(s) that the instructor is teaching,

• Link to their instructor website, if the instructor has a website.

Change the placeholder of “Add title” to “Add staff name” when adding a new Staff post.

Staff Template – 4 points

Create a template for showing all the staff custom post type items and a paragraph of text before them.

Display the items in the Faculty and Administrative terms separately on the template.

Display the following content for each staff member:

• Show their name,

• Show the short bio of the staff member,

• The course(s) the instructor is teaching (if Faculty),

• The instructor’s website URL (if Faculty).